requirements.txt
beautifulsoup4==4.13.4
jieba==0.42.1
numpy==2.2.6
pdfplumber==0.11.6
playwright==1.52.0
python_docx==1.1.2
rank_bm25==0.2.2
Requests==2.32.3
streamlit==1.45.1
torch==2.7.0
transformers==4.52.3
volcengine_python_sdk==3.0.4
datasets==3.6.0
httpx==0.28.1
pydantic==2.11.5
scikit-learn==1.6.1